moved subs count to its own file
